requirements.txt
# requirements.txt
openai
flask
python-dotenv
